requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
appnope==0.1.0            # via ipython
click==6.4                # via python-dotenv
coverage==4.0.3           # via pytest-cov
decorator==4.0.9          # via ipython, traitlets
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
Django==1.9.4             # via django-debug-toolbar, django-environ
Flake8==2.5.4
gnureadline==6.3.3        # via ipython
gunicorn==19.4.5
ipdb==0.9.0
ipython-genutils==0.1.0   # via traitlets
ipython==4.1.2
isort==4.2.5
mccabe==0.4.0             # via flake8
ordereddict==1.1          # via python-dotenv
path.py==8.1.2            # via pickleshare
pep8==1.7.0               # via flake8
pexpect==4.0.1            # via ipython
pickleshare==0.6          # via ipython
psycopg2==2.6.1
ptyprocess==0.5.1         # via pexpect
py==1.4.31                # via pytest
pyflakes==1.0.0           # via flake8
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-pythonpath==0.7
pytest==2.9.1             # via pytest-cov, pytest-django, pytest-pythonpath
python-dotenv==0.4.0
pytz==2016.3
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-environ, django-extensions
sqlparse==0.1.19          # via django-debug-toolbar
traitlets==4.2.1          # via ipython

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipdb, ipython
